Extract sortedMemos to shared utility
Remove duplicate sortedMemos function from MemosCrudReducer and
MemosLoadingReducer, replacing with shared internal utility.
