+++ feature/memos/presentation/src/commonMain/kotlin/space/be1ski/vibits/feature/memos/presentation/reducer/MemosLoadingReducer.kt
@@ -3,7 +3,6 @@ package space.be1ski.vibits.feature.memos.presentation.reducer
 import space.be1ski.vibits.core.elm.Reducer
 import space.be1ski.vibits.core.elm.reducer
 import space.be1ski.vibits.core.platform.mode.AppMode
-import space.be1ski.vibits.feature.memos.domain.model.Memo
 import space.be1ski.vibits.feature.memos.presentation.action.MemosAction
 import space.be1ski.vibits.feature.memos.presentation.effect.MemosEffect
 import space.be1ski.vibits.feature.memos.presentation.state.MemosState
@@ -87,24 +86,3 @@ internal val loadingReducer: Reducer<MemosAction.Loading, MemosState, MemosEffec
       }
     }
   }
-
-private fun sortedMemos(memos: List<Memo>): List<Memo> {
-  return memos.sortedByDescending { memo ->
-    // For habit tracking posts, use the tracked date instead of creation date
-    val trackingDate =
-      space.be1ski.vibits.feature.habits.domain.usecase
-        .parseDailyDateFromContent(memo.content)
-    if (trackingDate != null) {
-      // Convert LocalDate to epoch days, then to milliseconds
-      // toEpochDays returns days since 1970-01-01
-      trackingDate.toEpochDays() * MILLIS_PER_DAY
-    } else {
-      // For all other posts, use createTime or updateTime
-      memo.createTime?.toEpochMilliseconds()
-        ?: memo.updateTime?.toEpochMilliseconds()
-        ?: Long.MIN_VALUE
-    }
-  }
-}
-
-private const val MILLIS_PER_DAY = 86400000L
